refactor common module
