+++ common/src/main/kotlin/su/afk/kemonos/common/data/dto/ProfilePostsDto.kt
@@ -1,8 +1,8 @@
-package su.afk.kemonos.common.data.common
+package su.afk.kemonos.common.data.dto
 
 import com.google.gson.annotations.SerializedName
-import su.afk.kemonos.common.data.common.AttachmentDto.Companion.toDomain
-import su.afk.kemonos.common.data.common.FileDto.Companion.toDomain
+import su.afk.kemonos.common.data.dto.AttachmentDto.Companion.toDomain
+import su.afk.kemonos.common.data.dto.FileDto.Companion.toDomain
 import su.afk.kemonos.domain.models.PostDomain
 
 /** Общая модель для поиск и избранного */
