+++ common/src/main/kotlin/su/afk/kemonos/common/shared/ShareActions.kt
@@ -3,26 +3,8 @@ package su.afk.kemonos.common.shared
 import android.content.ClipData
 import android.content.ClipboardManager
 import android.content.Context
-import android.content.Intent
-import androidx.core.content.ContextCompat
 
 object ShareActions {
-
-    fun shareText(
-        context: Context,
-        text: String,
-        title: String = "Share",
-    ) {
-        val intent = Intent(Intent.ACTION_SEND).apply {
-            type = "text/plain"
-            putExtra(Intent.EXTRA_TEXT, text)
-        }
-        ContextCompat.startActivity(
-            context,
-            Intent.createChooser(intent, title).addFlags(Intent.FLAG_ACTIVITY_NEW_TASK),
-            null
-        )
-    }
 
     fun copyToClipboard(
         context: Context,
