+++ feature/creatorProfile/src/main/java/su/afk/kemonos/creatorProfile/data/api/CreatorProfileApi.kt
@@ -4,7 +4,7 @@ import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Path
 import retrofit2.http.Query
-import su.afk.kemonos.common.data.common.PostUnifiedDto
+import su.afk.kemonos.common.data.dto.PostUnifiedDto
 import su.afk.kemonos.creatorProfile.data.dto.profile.ProfileDto
 import su.afk.kemonos.creatorProfile.data.dto.profileAnnouncements.ProfileAnnouncementsDto
 import su.afk.kemonos.creatorProfile.data.dto.profileDms.DmDto
