+++ feature/creators/src/main/java/su/afk/kemonos/creators/data/dto/CreatorsDto.kt
@@ -1,4 +1,4 @@
-package su.afk.kemonos.common.data.creators
+package su.afk.kemonos.creators.data.dto
 
 import com.google.gson.annotations.SerializedName
 import su.afk.kemonos.domain.models.Creators
