+++ feature/posts-api/src/main/kotlin/su/afk/kemonos/posts/api/ICheckApiUseCase.kt
@@ -1,0 +1,7 @@
+package su.afk.kemonos.posts.api
+
+import su.afk.kemonos.posts.api.apiCheck.ApiCheckResult
+
+interface ICheckApiUseCase {
+    suspend operator fun invoke(): ApiCheckResult
+}
