+++ feature/posts-api/src/main/kotlin/su/afk/kemonos/posts/api/IPostsNavigator.kt
@@ -1,8 +1,7 @@
-package su.afk.kemonos.api
+package su.afk.kemonos.posts.api
 
 import androidx.navigation3.runtime.NavKey
 
 interface IPostsNavigator {
-
     fun getPostsDest(): NavKey
 }
