+++ feature/posts/src/main/java/su/afk/kemonos/posts/data/PostsRepository.kt
@@ -1,11 +1,11 @@
 package su.afk.kemonos.posts.data
 
-import su.afk.kemonos.api.domain.popular.PopularPosts
-import su.afk.kemonos.api.domain.tags.Tags
-import su.afk.kemonos.common.data.common.PostUnifiedDto.Companion.toDomain
+import su.afk.kemonos.common.data.dto.PostUnifiedDto.Companion.toDomain
 import su.afk.kemonos.domain.SelectedSite
 import su.afk.kemonos.domain.models.PostDomain
 import su.afk.kemonos.network.util.call
+import su.afk.kemonos.posts.api.popular.PopularPosts
+import su.afk.kemonos.posts.api.tags.Tags
 import su.afk.kemonos.posts.data.api.PostsApi
 import su.afk.kemonos.posts.data.dto.popular.response.PopularPostsDto.Companion.toDomain
 import su.afk.kemonos.posts.data.dto.random.RandomDto.Companion.toDomain
