+++ feature/posts/src/main/java/su/afk/kemonos/posts/data/dto/popular/response/PopularPostsDto.kt
@@ -1,12 +1,12 @@
 package su.afk.kemonos.posts.data.dto.popular.response
 
 import com.google.gson.annotations.SerializedName
-import su.afk.kemonos.api.domain.popular.PopularInfo
-import su.afk.kemonos.api.domain.popular.PopularNavigationDates
-import su.afk.kemonos.api.domain.popular.PopularPosts
-import su.afk.kemonos.api.domain.popular.PopularProps
-import su.afk.kemonos.common.data.common.PostUnifiedDto
-import su.afk.kemonos.common.data.common.PostUnifiedDto.Companion.toDomain
+import su.afk.kemonos.common.data.dto.PostUnifiedDto
+import su.afk.kemonos.common.data.dto.PostUnifiedDto.Companion.toDomain
+import su.afk.kemonos.posts.api.popular.PopularInfo
+import su.afk.kemonos.posts.api.popular.PopularNavigationDates
+import su.afk.kemonos.posts.api.popular.PopularPosts
+import su.afk.kemonos.posts.api.popular.PopularProps
 import su.afk.kemonos.posts.data.dto.popular.response.PopularInfoDto.Companion.toDomain
 import su.afk.kemonos.posts.data.dto.popular.response.PopularNavigationDatesDto.Companion.toDomain
 import su.afk.kemonos.posts.data.dto.popular.response.PopularPropsDto.Companion.toDomain
