+++ feature/posts/src/main/java/su/afk/kemonos/posts/domain/pagingPopular/GetPopularPostsUseCase.kt
@@ -4,9 +4,9 @@ import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.PagingData
 import kotlinx.coroutines.flow.Flow
-import su.afk.kemonos.api.domain.popular.PopularPosts
 import su.afk.kemonos.domain.SelectedSite
 import su.afk.kemonos.domain.models.PostDomain
+import su.afk.kemonos.posts.api.popular.PopularPosts
 import su.afk.kemonos.posts.data.PostsRepository
 import su.afk.kemonos.posts.domain.model.popular.Period
 import javax.inject.Inject
