+++ feature/posts/src/main/java/su/afk/kemonos/posts/domain/pagingPopular/PopularPostsPagingSource.kt
@@ -2,9 +2,9 @@ package su.afk.kemonos.posts.domain.pagingPopular
 
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
-import su.afk.kemonos.api.domain.popular.PopularPosts
 import su.afk.kemonos.domain.SelectedSite
 import su.afk.kemonos.domain.models.PostDomain
+import su.afk.kemonos.posts.api.popular.PopularPosts
 import su.afk.kemonos.posts.data.PostsRepository
 import su.afk.kemonos.posts.domain.model.popular.Period
