+++ feature/posts/src/main/java/su/afk/kemonos/posts/domain/usecase/GetAllTagsUseCase.kt
@@ -1,7 +1,7 @@
 package su.afk.kemonos.posts.domain.usecase
 
-import su.afk.kemonos.api.domain.tags.Tags
 import su.afk.kemonos.domain.SelectedSite
+import su.afk.kemonos.posts.api.tags.Tags
 import su.afk.kemonos.posts.data.PostsRepository
 import javax.inject.Inject
