+++ feature/posts/src/main/java/su/afk/kemonos/posts/navigation/PostsNavigator.kt
@@ -1,7 +1,7 @@
 package su.afk.kemonos.posts.navigation
 
 import androidx.navigation3.runtime.NavKey
-import su.afk.kemonos.api.IPostsNavigator
+import su.afk.kemonos.posts.api.IPostsNavigator
 import javax.inject.Inject
 
 class PostsNavigator @Inject constructor() : IPostsNavigator {
