+++ feature/posts/src/main/java/su/afk/kemonos/posts/presenter/pagePopularPosts/PopularPostsState.kt
@@ -3,9 +3,9 @@ package su.afk.kemonos.posts.presenter.pagePopularPosts
 import androidx.paging.PagingData
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.emptyFlow
-import su.afk.kemonos.api.domain.popular.PopularInfo
-import su.afk.kemonos.api.domain.popular.PopularProps
 import su.afk.kemonos.domain.models.PostDomain
+import su.afk.kemonos.posts.api.popular.PopularInfo
+import su.afk.kemonos.posts.api.popular.PopularProps
 import su.afk.kemonos.posts.domain.model.popular.Period
 
 internal data class PopularPostsState(
