+++ feature/posts/src/main/java/su/afk/kemonos/posts/presenter/pageTags/TagsPageState.kt
@@ -1,6 +1,6 @@
 package su.afk.kemonos.posts.presenter.pageTags
 
-import su.afk.kemonos.api.domain.tags.Tags
+import su.afk.kemonos.posts.api.tags.Tags
 
 internal data class TagsPageState(
     val loading: Boolean = false,
