+++ feature/posts/src/main/java/su/afk/kemonos/posts/presenter/pageTags/TagsPageViewModel.kt
@@ -2,13 +2,13 @@ package su.afk.kemonos.posts.presenter.pageTags
 
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
-import su.afk.kemonos.api.domain.tags.Tags
 import su.afk.kemonos.common.error.IErrorHandlerUseCase
 import su.afk.kemonos.common.error.storage.RetryStorage
 import su.afk.kemonos.common.presenter.changeSite.SiteAwareBaseViewModel
 import su.afk.kemonos.domain.SelectedSite
 import su.afk.kemonos.navigation.NavigationManager
 import su.afk.kemonos.navigation.NavigationStorage
+import su.afk.kemonos.posts.api.tags.Tags
 import su.afk.kemonos.posts.domain.usecase.GetAllTagsUseCase
 import su.afk.kemonos.posts.navigation.PostsDest
 import su.afk.kemonos.posts.presenter.util.Const.KEY_SELECTED_TAG
