+++ feature/profile/src/main/java/su/afk/kemonos/profile/data/api/FavoritesApi.kt
@@ -3,7 +3,7 @@ package su.afk.kemonos.profile.data.api
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
-import su.afk.kemonos.common.data.common.PostUnifiedDto
+import su.afk.kemonos.common.data.dto.PostUnifiedDto
 import su.afk.kemonos.network.auth.AuthCookie
 import su.afk.kemonos.network.creators.HeaderText
 import su.afk.kemonos.profile.data.dto.favorites.artist.FavoriteArtistDto
