+++ storage-api/src/main/kotlin/su/afk/kemonos/storage/api/popular/IPopularPostsCacheUseCase.kt
@@ -1,7 +1,7 @@
 package su.afk.kemonos.storage.api.popular
 
-import su.afk.kemonos.api.domain.popular.PopularPosts
 import su.afk.kemonos.domain.SelectedSite
+import su.afk.kemonos.posts.api.popular.PopularPosts
 
 interface IPopularPostsCacheUseCase {
     suspend fun getFreshOrNull(site: SelectedSite, queryKey: String, offset: Int): PopularPosts?
