+++ storage-api/src/main/kotlin/su/afk/kemonos/storage/api/tags/IStoreTagsUseCase.kt
@@ -1,7 +1,7 @@
 package su.afk.kemonos.storage.api.tags
 
-import su.afk.kemonos.api.domain.tags.Tags
 import su.afk.kemonos.domain.SelectedSite
+import su.afk.kemonos.posts.api.tags.Tags
 
 interface IStoreTagsUseCase {
     suspend fun getAll(site: SelectedSite): List<Tags>
